sslit out overlay handling
