+++ src/utils/highlighter.ts
@@ -1,39 +1,49 @@
 import browser from './browser-polyfill';
-import { throttle } from './throttle';
 import { getElementXPath, getElementByXPath } from './dom-utils';
+import {
+	handleMouseUp,
+	handleMouseMove,
+	removeHoverOverlay,
+	updateHighlightListeners,
+	createHighlightOverlayGroup,
+	removeExistingHighlights
+} from './highlighter-overlays';
 
-let isHighlighterMode = false;
-let highlights: AnyHighlightData[] = [];
-let hoverOverlay: HTMLElement | null = null;
-let isApplyingHighlights = false;
+export type AnyHighlightData = TextHighlightData | ElementHighlightData | ComplexHighlightData;
+
+export let isHighlighterMode = false;
+export let highlights: AnyHighlightData[] = [];
+export let isApplyingHighlights = false;
 let lastAppliedHighlights: string = '';
 let originalLinkClickHandlers: WeakMap<HTMLElement, (event: MouseEvent) => void> = new WeakMap();
 
-interface HighlightData {
+export interface HighlightData {
 	xpath: string;
 	content: string;
 	id: string;
 }
 
-interface TextHighlightData extends HighlightData {
+export interface TextHighlightData extends HighlightData {
 	type: 'text';
 	startOffset: number;
 	endOffset: number;
 }
 
-interface ElementHighlightData extends HighlightData {
+export interface ElementHighlightData extends HighlightData {
 	type: 'element';
 }
 
-interface ComplexHighlightData extends HighlightData {
+export interface ComplexHighlightData extends HighlightData {
 	type: 'complex';
 }
 
-type AnyHighlightData = TextHighlightData | ElementHighlightData | ComplexHighlightData;
-
-interface StoredData {
+export interface StoredData {
 	highlights: AnyHighlightData[];
 	url: string;
+}
+
+export function updateHighlights(newHighlights: AnyHighlightData[]) {
+	highlights = newHighlights;
 }
 
 // Toggle highlighter mode on or off
@@ -80,40 +90,8 @@ function enableLinkClicks() {
 	});
 }
 
-// Handle mouse up events for highlighting
-function handleMouseUp(event: MouseEvent) {
-	if (!isHighlighterMode) return;
-	const selection = window.getSelection();
-	if (selection && !selection.isCollapsed) {
-		handleTextSelection(selection);
-	} else {
-		const target = event.target as Element;
-		if (target.classList.contains('obsidian-highlight-overlay')) {
-			removeHighlightByElement(target);
-		} else if (!isIgnoredElement(target)) {
-			highlightElement(target);
-		}
-	}
-}
-
-// Handles mouse move events for hover effects
-function handleMouseMove(event: MouseEvent) {
-	if (!isHighlighterMode) return;
-	const target = event.target as Element;
-	if (!isIgnoredElement(target)) {
-		createHoverOverlay(target);
-	} else {
-		removeHoverOverlay();
-	}
-}
-
-// Check if an element should be ignored for highlighting
-function isIgnoredElement(element: Element): boolean {
-	return element.tagName.toLowerCase() === 'html' || element.tagName.toLowerCase() === 'body';
-}
-
 // Highlight an entire element
-function highlightElement(element: Element) {
+export function highlightElement(element: Element) {
 	const xpath = getElementXPath(element);
 	const content = element.outerHTML;
 	const isBlockElement = window.getComputedStyle(element).display === 'block';
@@ -128,7 +106,7 @@ function highlightElement(element: Element) {
 }
 
 // Handle text selection for highlighting
-function handleTextSelection(selection: Selection) {
+export function handleTextSelection(selection: Selection) {
 	const range = selection.getRangeAt(0);
 	const highlightRanges = getHighlightRanges(range);
 	highlightRanges.forEach(hr => addHighlight(hr));
@@ -234,7 +212,7 @@ function addHighlight(highlight: AnyHighlightData) {
 }
 
 // Sort highlights based on their vertical position
-function sortHighlights() {
+export function sortHighlights() {
 	highlights.sort((a, b) => {
 		const elementA = getElementByXPath(a.xpath);
 		const elementB = getElementByXPath(b.xpath);
@@ -243,6 +221,11 @@ function sortHighlights() {
 		}
 		return 0;
 	});
+}
+
+// Get the vertical position of an element
+function getElementVerticalPosition(element: Element): number {
+	return element.getBoundingClientRect().top + window.scrollY;
 }
 
 // Check if two highlights overlap
@@ -369,7 +352,7 @@ function findCommonAncestor(element1: Element, element2: Element): Element {
 	return document.body; // Fallback to body if no common ancestor found
 }
 
-// Getsall parent elements of a given element
+// Get all parent elements of a given element
 function getParents(element: Element): Element[] {
 	const parents: Element[] = [];
 	let currentElement: Element | null = element;
@@ -381,16 +364,6 @@ function getParents(element: Element): Element[] {
 
 	parents.unshift(document.body);
 	return parents;
-}
-
-// Update event listeners for highlight overlays
-export function updateHighlightListeners() {
-	document.querySelectorAll('.obsidian-highlight-overlay').forEach(highlight => {
-		highlight.removeEventListener('click', removeHighlightByEvent);
-		if (isHighlighterMode) {
-			highlight.addEventListener('click', removeHighlightByEvent);
-		}
-	});
 }
 
 // Save highlights to browser storage
@@ -436,171 +409,6 @@ export function applyHighlights() {
 	notifyHighlightsUpdated();
 }
 
-// Find a text node at a given offset within an element
-function findTextNodeAtOffset(element: Element, offset: number): { node: Node, offset: number } | null {
-	let currentOffset = 0;
-	const treeWalker = document.createTreeWalker(element, NodeFilter.SHOW_TEXT);
-	
-	let node: Node | null = treeWalker.currentNode;
-	while (node) {
-		const nodeLength = node.textContent?.length || 0;
-		if (currentOffset + nodeLength >= offset) {
-			return { node, offset: offset - currentOffset };
-		}
-		currentOffset += nodeLength;
-		node = treeWalker.nextNode();
-	}
-	
-	// If we couldn't find the exact offset, return the last text node and its length
-	const lastNode = treeWalker.lastChild();
-	if (lastNode) {
-		return { node: lastNode, offset: lastNode.textContent?.length || 0 };
-	}
-	
-	return null;
-}
-
-// Calculate the average line height of a set of rectangles
-function calculateAverageLineHeight(rects: DOMRectList): number {
-	const heights = Array.from(rects).map(rect => rect.height);
-	const sum = heights.reduce((a, b) => a + b, 0);
-	return sum / heights.length;
-}
-
-// Create a group for elements that are part of the same highlight
-function createHighlightOverlayGroup(target: Element, index: number, highlight: AnyHighlightData) {
-	let container = document.querySelector(`.obsidian-highlight-container[data-highlight-index="${index}"]`) as HTMLElement;
-	
-	if (!container) {
-		container = document.createElement('div');
-		container.className = 'obsidian-highlight-container';
-		container.dataset.highlightIndex = index.toString();
-		container.addEventListener('click', handleHighlightClick);
-		document.body.appendChild(container);
-	}
-	
-	planHighlightOverlayRects(target, container, highlight);
-}
-
-// Plan out the overlay rectangles depending on the type of highlight, i.e. individual lines of text or entire elements
-function planHighlightOverlayRects(target: Element, container: HTMLElement, highlight: AnyHighlightData) {
-	const existingOverlays = Array.from(document.querySelectorAll('.obsidian-highlight-overlay'));
-	if (highlight.type === 'complex' || highlight.type === 'element') {
-		const rect = target.getBoundingClientRect();
-		mergeHighlightOverlayRects([rect], container, highlight.content, existingOverlays, false);
-	} else if (highlight.type === 'text') {
-		try {
-			const range = document.createRange();
-			const startNode = findTextNodeAtOffset(target, highlight.startOffset);
-			const endNode = findTextNodeAtOffset(target, highlight.endOffset);
-			
-			if (startNode && endNode) {
-				range.setStart(startNode.node, startNode.offset);
-				range.setEnd(endNode.node, endNode.offset);
-				
-				const rects = range.getClientRects();
-				
-				const averageLineHeight = calculateAverageLineHeight(rects);
-				const textRects = Array.from(rects).filter(rect => rect.height <= averageLineHeight * 1.5);
-				const complexRects = Array.from(rects).filter(rect => rect.height > averageLineHeight * 1.5);
-				
-				// Only create overlays for complex rects if there are no text rects
-				if (textRects.length > 0) {
-					mergeHighlightOverlayRects(textRects, container, highlight.content, existingOverlays, true);
-				} else {
-					mergeHighlightOverlayRects(complexRects, container, highlight.content, existingOverlays, false);
-				}
-			} else {
-				console.warn('Could not find start or end node for text highlight, falling back to element highlight');
-				const rect = target.getBoundingClientRect();
-				mergeHighlightOverlayRects([rect], container, highlight.content, existingOverlays, false);
-			}
-		} catch (error) {
-			console.error('Error creating text highlight, falling back to element highlight:', error);
-			const rect = target.getBoundingClientRect();
-			mergeHighlightOverlayRects([rect], container, highlight.content, existingOverlays, false);
-		}
-	}
-}
-
-// Merge a set of rectangles, to avoid adjacent and overlapping highlights where possible
-function mergeHighlightOverlayRects(rects: DOMRect[], container: HTMLElement, content: string, existingOverlays: Element[], isText: boolean = false) {
-	let overlaysCreated = 0;
-
-	let mergedRects: DOMRect[] = [];
-	let currentRect: DOMRect | null = null;
-
-	for (let i = 0; i < rects.length; i++) {
-		const rect = rects[i];
-		if (!currentRect) {
-			currentRect = new DOMRect(rect.x, rect.y, rect.width, rect.height);
-		} else if (Math.abs(rect.y - currentRect.y) < 1 && Math.abs(rect.height - currentRect.height) < 1) {
-			// Merge adjacent rects with the same height and y-position
-			currentRect.width = rect.right - currentRect.left;
-		} else {
-			mergedRects.push(currentRect);
-			currentRect = new DOMRect(rect.x, rect.y, rect.width, rect.height);
-		}
-	}
-	if (currentRect) {
-		mergedRects.push(currentRect);
-	}
-
-	for (const rect of mergedRects) {
-		const isDuplicate = existingOverlays.some(overlay => {
-			const overlayRect = overlay.getBoundingClientRect();
-			const duplicate = (
-				Math.abs(rect.left - overlayRect.left) < 1 &&
-				Math.abs(rect.top - overlayRect.top) < 1 &&
-				Math.abs(rect.width - overlayRect.width) < 1 &&
-				Math.abs(rect.height - overlayRect.height) < 1
-			);
-			return duplicate;
-		});
-
-		if (!isDuplicate) {
-			createHighlightOverlayElement(rect, container, content, isText);
-			overlaysCreated++;
-		}
-	}
-}
-
-// Create an overlay element
-function createHighlightOverlayElement(rect: DOMRect, container: HTMLElement, content: string, isText: boolean = false) {
-	const overlay = document.createElement('div');
-	overlay.className = 'obsidian-highlight-overlay';
-	
-	overlay.style.position = 'absolute';
-	overlay.style.left = `${rect.left + window.scrollX}px`;
-	overlay.style.top = `${rect.top + window.scrollY}px`;
-	overlay.style.width = `${rect.width}px`;
-	overlay.style.height = `${rect.height}px`;
-	
-	overlay.setAttribute('title', content);
-	
-	container.appendChild(overlay);
-}
-
-// Handle click events on highlight overlays
-function handleHighlightClick(event: MouseEvent) {
-	event.stopPropagation();
-	const container = event.currentTarget as HTMLElement;
-	removeHighlightByElement(container);
-}
-
-// Remove a highlight based on its container element
-function removeHighlightByElement(container: Element) {
-	const index = container.getAttribute('data-highlight-index');
-	if (index !== null) {
-		const highlightToRemove = highlights[parseInt(index)];
-		highlights = highlights.filter(h => h.id !== highlightToRemove.id);
-		container.remove();
-		sortHighlights();
-		applyHighlights();
-		saveHighlights();
-	}
-}
-
 // Notify that highlights have been updated
 function notifyHighlightsUpdated() {
 	browser.runtime.sendMessage({ action: "highlightsUpdated" });
@@ -623,93 +431,6 @@ export function loadHighlights() {
 	});
 }
 
-// Update positions of all highlight overlays
-function updateHighlightOverlayPositions() {
-	document.querySelectorAll('.obsidian-highlight-container').forEach((container: Element) => {
-		const index = container.getAttribute('data-highlight-index');
-		if (index !== null) {
-			const highlight = highlights[parseInt(index)];
-			const target = getElementByXPath(highlight.xpath);
-			if (target) {
-				updateHighlightOverlay(target, container as HTMLElement, highlight);
-			}
-		}
-	});
-}
-
-// Update a highlight overlay
-function updateHighlightOverlay(target: Element, container: HTMLElement, highlight: AnyHighlightData) {
-	container.innerHTML = ''; // Clear existing overlay elements
-	
-	planHighlightOverlayRects(target, container, highlight);
-}
-
-const throttledUpdateHighlights = throttle(() => {
-	if (!isApplyingHighlights) {
-		updateHighlightOverlayPositions();
-	}
-}, 100);
-
-window.addEventListener('resize', throttledUpdateHighlights);
-window.addEventListener('scroll', throttledUpdateHighlights);
-
-const observer = new MutationObserver((mutations) => {
-	if (!isApplyingHighlights) {
-		const shouldUpdate = mutations.some(mutation => 
-			mutation.type === 'childList' || 
-			(mutation.type === 'attributes' && 
-			 (mutation.attributeName === 'style' || mutation.attributeName === 'class'))
-		);
-		if (shouldUpdate) {
-			throttledUpdateHighlights();
-		}
-	}
-});
-observer.observe(document.body, { 
-	childList: true, 
-	subtree: true, 
-	attributes: true,
-	attributeFilter: ['style', 'class'],
-	characterData: false
-});
-
-// Create the hover overlay used to indicate which element will be highlighted
-function createHoverOverlay(target: Element) {
-	removeHoverOverlay();
-	
-	hoverOverlay = document.createElement('div');
-	hoverOverlay.className = 'obsidian-highlight-hover-overlay';
-	
-	const rect = target.getBoundingClientRect();
-	
-	hoverOverlay.style.position = 'absolute';
-	hoverOverlay.style.left = `${rect.left + window.scrollX - 2}px`;
-	hoverOverlay.style.top = `${rect.top + window.scrollY - 2}px`;
-	hoverOverlay.style.width = `${rect.width + 4}px`;
-	hoverOverlay.style.height = `${rect.height + 4}px`;
-	
-	document.body.appendChild(hoverOverlay);
-}
-
-// Removes the hover overlay
-function removeHoverOverlay() {
-	if (hoverOverlay) {
-		hoverOverlay.remove();
-		hoverOverlay = null;
-	}
-}
-
-// Remove a highlight in response to an event
-function removeHighlightByEvent(event: Event) {
-	const overlay = event.currentTarget as HTMLElement;
-	removeHighlightByElement(overlay);
-}
-
-// Remove all existing highlight overlays from the page
-export function removeExistingHighlights() {
-	document.querySelectorAll('.obsidian-highlight-container').forEach(el => el.remove());
-}
-
 // Clear all highlights from the page and storage
 export function clearHighlights() {
 	const url = window.location.href;
@@ -718,9 +439,4 @@ export function clearHighlights() {
 		removeExistingHighlights();
 		console.log('Highlights cleared for:', url);
 	});
-}
-
-// Get the vertical position of an element
-function getElementVerticalPosition(element: Element): number {
-	return element.getBoundingClientRect().top + window.scrollY;
 }
